chore: fix e2e
