+++ .github/workflows/e2e.yml
@@ -65,7 +65,10 @@ jobs:
         env: 
           NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
 
-      - run: npm install
+      - name: Install Packages
+        run: npm install
+        env: 
+          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
 
       - name: Run E2E Test on iOS
         run: npm run e2e:ios
@@ -96,7 +99,10 @@ jobs:
         env: 
           NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
 
-      - run: npm install
+      - name: Install Packages
+        run: npm install
+        env: 
+          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
 
       - name: AVD cache
         uses: actions/cache@v2
