+++ google-maps/README.md
@@ -55,6 +55,7 @@ destroy(args: DestroyMapArgs) => Promise<void>
 
 | Prop              | Type                                                        |
 | ----------------- | ----------------------------------------------------------- |
+| **`id`**          | <code>string</code>                                         |
 | **`config`**      | <code><a href="#googlemapconfig">GoogleMapConfig</a></code> |
 | **`forceCreate`** | <code>boolean</code>                                        |
 
@@ -81,5 +82,9 @@ destroy(args: DestroyMapArgs) => Promise<void>
 
 
 #### DestroyMapArgs
+
+| Prop     | Type                |
+| -------- | ------------------- |
+| **`id`** | <code>string</code> |
 
 </docgen-api>
