+++ google-maps/e2e-tests/src/pages/Home.tsx
@@ -2,7 +2,7 @@ import BaseTestingPage from '../components/BaseTestingPage';
 
 const Home: React.FC = () => {
   return (
-    <BaseTestingPage pageTitle="IV Test App">
+    <BaseTestingPage pageTitle="Test App">
       <h3 className="ion-padding" style={{ textAlign: 'center' }}>
         Choose a testing page from the menu to get started!
       </h3>
