+++ google-maps/src/definitions.ts
@@ -1,0 +1,14 @@
+export interface LatLng {
+  lat: number;
+  lng: number;
+}
+
+export interface GoogleMapConfig {
+  width: number;
+  height: number;
+  x: number;
+  y: number;
+  center: LatLng;
+  zoom: number;
+  androidLiteMode: boolean;
+}
